requirements.txt
line-bot-sdk==3.7.0
flask==3.0.0
python-dotenv==1.0.0
python-json-logger==2.0.7
pyngrok==7.1.5
gunicorn==21.2.0
requests==2.31.0
pytest==8.1.1
black==24.3.0
flake8==7.0.0 